chore: extract useMultisigInfo
